Added functions multiply and divide. and added float numbers.

+++ src/functions/addition.js
@@ -1,1 +1,1 @@
-export default function addition (n1, n2) {return parseInt(n1)+parseInt(n2)};
+export default function addition (n1, n2) {return parseFloat(n1)+parseFloat(n2)};
