+++ src/functions/subtract.js
@@ -1,1 +1,1 @@
-export default function subtract (n1, n2) {return parseInt(n1)-parseInt(n2)};
+export default function subtract (n1, n2) {return parseFloat(n1)-parseFloat(n2)};
